Reference

Contents

Reference

Contents

Index

Index

ShiftedProximalOperators.RootNormLhalf

ShiftedProximalOperators.ShiftedProximableFunction

ProximalOperators.prox

ProximalOperators.prox!

ShiftedProximalOperators.set_radius!

ShiftedProximalOperators.shift!

ShiftedProximalOperators.shifted

ShiftedProximalOperators.RootNormLhalf — Type

$L_1/2^(1/2)$ pseudo-norm RootNormLhalf(λ=1) Returns the function

\[f(x) = λ\cdot \sum |x|\^{1/2}\]

for a nonnegative parameter λ.

ShiftedProximalOperators.ShiftedProximableFunction — Type

Abstract type for shifted proximable functions.

ProximalOperators.prox! — Function

prox!(y, ψ, q, σ)

Evaluate the proximal operator of a shifted regularizer, i.e, return a solution s of

minimize{s}  ½ σ⁻¹ ‖s - q‖₂² + ψ(s),

where

ψ is a ShiftedProximableFunction representing a model of h(x + s) and possibly including the indicator of a trust region;

q is the vector where the shifted proximal operator should be evaluated;

σ is a positive regularization parameter.

The solution is stored in the input vector y an y is returned.

ProximalOperators.prox — Method

prox(ψ, q, σ)

See the documentation of prox!. In this form, the solution is stored in ψ's internal storage and a reference is returned.

ShiftedProximalOperators.set_radius! — Method

set_radius!(ψ, Δ)

Set the trust-region radius of a shifted proximable function to Δ. This method updates the indicator of the trust region that is part of ψ.

ShiftedProximalOperators.shift! — Method

shift!(ψ, x)

Update the shift of a shifted proximable function.

ShiftedProximalOperators.shifted — Function

shifted(h, x)
shifted(h, x, Δ, χ)

Construct a shifted proximable function from a proximable function or from another shifted proximable function.

If h is a ProximableFunction, including a ShiftedProximableFunction, the form shifted(h, x) returns a ShiftedProximableFunction ψ such that ψ(s) == h(x + s). Subsequently, prox may be called on ψ. The first form applies when h is a ShiftedProximableFunction and can be used to shift an already-shifted proximable function.

The form shifted(h, x, Δ, χ) returns a ShiftedProximableFunction ψ such that ψ(s) == h(x + s) + Ind({‖s‖ ≤ Δ}), where Ind(.) represents the indicator of a set, in this case the indicator of a ball of radius Δ, in which the norm is defined by χ.

Arguments

h::ProximableFunction (including a ShiftedProximableFunction)

x::AbstractVector

Δ::Real

χ::ProximableFunction.

The currently supported combinations are:

h::IndBallL0 and χ::Conjugate{IndBallL1} (i.e., χ is the Inf-norm)

h::NormL0 and χ::Conjugate{IndBallL1} (i.e., χ is the Inf-norm)

h::NormL1 and χ::Conjugate{IndBallL1} (i.e., χ is the Inf-norm)

h::NormL1 and χ::NormL2.

If h is a shifted proximable function obtained from a previous call to shifted(), only the form shifted(h, x) is supported. If applicable, the resulting shifted proximable function is associated with the same Δ and χ as h.

See the documentation of ProximalOperators.jl for more information.