Reference

Contents

Reference

Contents

Index

Index

ShiftedProximalOperators.Cappedl1

ShiftedProximalOperators.GroupNormL2

ShiftedProximalOperators.Nuclearnorm

ShiftedProximalOperators.Rank

ShiftedProximalOperators.RootNormLhalf

ShiftedProximalOperators.ShiftedProximableFunction

ProximalCore.prox

ProximalCore.prox!

ShiftedProximalOperators.iprox

ShiftedProximalOperators.iprox!

ShiftedProximalOperators.iprox_zero

ShiftedProximalOperators.prox_zero

ShiftedProximalOperators.set_bounds!

ShiftedProximalOperators.set_radius!

ShiftedProximalOperators.shift!

ShiftedProximalOperators.shifted

ShiftedProximalOperators.Cappedl1 — Type

Cappedl1(λ, θ)

Returns the capped L1 approximation to the rank function

\[f(x) = \lambda \min(θ, σ(X))\]

for nonnegative parameters $\lambda$ and $\theta$, and a vector $x$, where $x = \text{vec}(X)$.

ShiftedProximalOperators.GroupNormL2 — Type

GroupNormL2(λ = 1, idx = [:])

Returns the group $\ell_2$-norm operator

\[f(x) = \sum_i \lambda_i \|x_{[i]}\|_2^{1/2}\]

for groups $x_{[i]}$ and nonnegative weights $\lambda_i$. The group $\ell_2$-norm operator reduces to the $\ell_2$-norm if only one group is defined (the default).

ShiftedProximalOperators.Nuclearnorm — Type

Nuclearnorm(λ)

Returns the nuclear norm

\[f(x) = \lambda \|X\|_*\]

for a nonnegative parameter $\lambda$ and a vector $x$, where $x = \text{vec}(X)$.

ShiftedProximalOperators.Rank — Type

Rank(λ)

Returns the rank

\[f(x) = \lambda \text{rank}(X)\]

for a nonnegative parameter $\lambda$ and a vector $x$, where $x = \text{vec}(X)$.

ShiftedProximalOperators.RootNormLhalf — Type

RootNormLhalf(λ=1)

Returns the $\ell_{1/2}^{1/2}$ pseudo-norm operator

\[f(x) = λ \sum |x|^{1/2}\]

where $\lambda > 0$.

ShiftedProximalOperators.ShiftedProximableFunction — Type

Abstract type for shifted proximable functions.

ProximalCore.prox! — Function

prox!(y, ψ, q, σ)

Evaluate the proximal operator of a shifted regularizer, i.e, return a solution s of

minimize{s}  ½ σ⁻¹ ‖s - q‖₂² + ψ(s),

where

ψ is a ShiftedProximableFunction representing a model of h(x + s) and possibly including the indicator of a trust region;

q is the vector where the shifted proximal operator should be evaluated;

σ is a positive regularization parameter.

The solution is stored in the input vector y an y is returned.

ProximalCore.prox — Method

prox(ψ, q, σ)

See the documentation of prox!. In this form, the solution is stored in ψ's internal storage and a reference is returned.

ShiftedProximalOperators.iprox! — Function

iprox!(y, ψ, g, d)

Evaluate the indefinite proximal operator of a separable box shifted regularizer, i.e, return a solution y of

minimize{y}  ½ yᵀDy + gᵀy + ψ(y)

where

ψ is a ShiftedProximableFunction representing a model of the sum of a separable function h(x + s) and the indicator of a trust region;

g is a vector;

D = Diagonal(d) where d is a vector.

The solution is stored in the input vector y an y is returned.

ShiftedProximalOperators.iprox — Method

iprox(ψ, g, d)

See the documentation of iprox!. In this form, the solution is stored in ψ's internal storage and a reference is returned.

ShiftedProximalOperators.iprox_zero — Method

iprox_zero(d, g, l, u)

Return the solution of

min ½ d y² + g y subject to l ≤ y ≤ u

for any d, g, assuming that 0 ∈ [l, u] if d < eps() and g = 0. This problem occurs when computing the iprox with respect to a separable nonsmooth term along a variable that is not part of those to which the nonsmooth term is applied.

ShiftedProximalOperators.prox_zero — Method

prox_zero(q, l, u)

Return the solution of

min ½ σ⁻¹ (y - q)² subject to l ≤ y ≤ u

for any σ > 0. This problem occurs when computing the prox with respect to a separable nonsmooth term along a variable that is not part of those to which the nonsmooth term is applied.

ShiftedProximalOperators.set_bounds! — Method

set_bounds!(ψ, l, u)

Set the lower and upper bound of the box to l and u, respectively. l and u can be scalars or vectors.

ShiftedProximalOperators.set_radius! — Method

set_radius!(ψ, Δ)

Set the trust-region radius of a shifted proximable function to Δ. This method updates the indicator of the trust region that is part of ψ.

ShiftedProximalOperators.shift! — Method

shift!(ψ, x)

Update the shift of a shifted proximable function.

ShiftedProximalOperators.shifted — Function

shifted(h, x)
shifted(h, x, Δ, ρ)

Construct a shifted proximable function from a proximable function or from another shifted proximable function.

If h is a ProximableFunction, including a ShiftedProximableFunction, the form shifted(h, x) returns a ShiftedProximableFunction ψ such that ψ(s) == h(x + s). Subsequently, prox may be called on ψ. The first form applies when h is a ShiftedProximableFunction and can be used to shift an already-shifted proximable function.

The form shifted(h, x, Δ, ρ) returns a ShiftedProximableFunction ψ such that ψ(s) == h(x + s) + Ind({‖s‖ ≤ Δ}), where Ind(.) represents the indicator of a set, in this case the indicator of a ball of radius Δ, in which the norm is defined by ρ.

Arguments

h::ProximableFunction (including a ShiftedProximableFunction)

x::AbstractVector

Δ::Real

ρ::ProximableFunction.

Only certain combinations of h and ρ are supported; those for which the analytical form of the proximal operator is known.

If h is a shifted proximable function obtained from a previous call to shifted(), only the form shifted(h, x) is supported. If applicable, the resulting shifted proximable function is associated with the same Δ and ρ as h.

See the documentation of ProximalOperators.jl for more information.